sunlight: RX/TX driver for sending and receiving messages
Simple packet_event_loop to first listen for incoming packages in the
current channel and if there's none and the outgoing packet queue is not
empty we switch to transmitting mode and send our packet. The packet
might be lost since the other device might not be listening. This means
that we need to check for returning message that'll acknowledge our
packet.

sky: Simple fixed size queue datastructure.